+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   annotation to force IntelliJ IDEA to treat annotate content as css rules for code formatting, completion
   and validation purposes. Use this new annotation to annotate relevant constants and parameters in
   the `ConsoleSpyEventProcessor` class.
+* **\[extras\]** Enhance the `ConsoleSpyEventProcessor` class so that javascript native objects passed as parameters
+  to actions are formatted using `JSON.stringify` so that they produce more human friendly messages.
 
 ## [v0.03](https://github.com/realityforge/arez/tree/v0.03) (2017-10-02)
 [Full Changelog](https://github.com/realityforge/arez/compare/v0.02...v0.03)